+++ android/src/main/kotlin/com/invisiblewrench/fluttermidicommand/FlutterMidiCommandPlugin.kt
@@ -582,13 +582,13 @@ class FlutterMidiCommandPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
         if (data.size > 0) {
           for (i in 0 until data.size) {
             var midiByte: Byte = data[i]
-            var midiInt = midiByte.toInt()
+            var midiInt = midiByte.toInt() and 0xFF
 
 //          Log.d("FlutterMIDICommand", "parserState $parserState byte $midiByte")
 
             when (parserState) {
               PARSER_STATE.HEADER -> {
-                if (midiInt and 0xFF == 0xF0) {
+                if (midiInt == 0xF0) {
                   parserState = PARSER_STATE.SYSEX
                   sysExBuffer.clear()
                   sysExBuffer.add(midiByte)
@@ -600,6 +600,7 @@ class FlutterMidiCommandPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
                   midiBuffer.clear()
                   midiBuffer.add(midiByte)
                   parserState = PARSER_STATE.PARAMS
+                  finalizeMessageIfComplete(timestamp)
                 } else {
                   // in header state but no status byte, do running status
                   midiBuffer.clear()
@@ -611,7 +612,7 @@ class FlutterMidiCommandPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
               }
 
               PARSER_STATE.SYSEX -> {
-                if (midiInt and 0xFF == 0xF0) {
+                if (midiInt == 0xF0) {
                   // Android can skip SysEx end bytes, when more sysex messages are coming in succession.
                   // in an attempt to save the situation, add an end byte to the current buffer and start a new one.
                   sysExBuffer.add(0xF7.toByte())
@@ -626,7 +627,7 @@ class FlutterMidiCommandPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
                   sysExBuffer.clear();
                 }
                 sysExBuffer.add(midiByte)
-                if (midiInt and 0xFF == 0xF7) {
+                if (midiInt == 0xF7) {
                   // Sysex complete
 //                Log.d("FlutterMIDICommand", "sysex complete $sysExBuffer")
                   stream.send(
@@ -664,11 +665,12 @@ class FlutterMidiCommandPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
       when (type) {
         0xF6, 0xF8, 0xFA, 0xFB, 0xFC, 0xFF, 0xFE -> return 1
         0xF1, 0xF3 -> return 2
+        0xF2 -> return 3
       }
 
       when (midiType) {
         0xC0, 0xD0 -> return 2
-        0xF2, 0x80, 0x90, 0xA0, 0xB0, 0xE0 -> return 3
+        0x80, 0x90, 0xA0, 0xB0, 0xE0 -> return 3
       }
       return 0
     }
@@ -801,7 +803,7 @@ class FlutterMidiCommandPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
           this.receiver?.send(data, 0, data.size, timestamp)
 
       } else {
-        Log.d("FlutterMIDICommand", "Send to input port ${this.inputPort}")
+//        Log.d("FlutterMIDICommand", "Send to input port ${this.inputPort}")
         this.inputPort?.send(data, 0, data.count(), if (timestamp is Long) timestamp else 0)
       }
     }