+++ lib/flutter_midi_command_messages.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 import 'flutter_midi_command.dart';
 
-enum MessageType { CC, PC, NoteOn, NoteOff, NRPN, SYSEX, Beat }
+enum MessageType { CC, PC, NoteOn, NoteOff, NRPN, SYSEX, Beat, PolyAT, AT, PitchBend }
 
 /// Base class for MIDI message types
 class MidiMessage {
@@ -206,6 +206,63 @@ class NRPNHexMessage extends MidiMessage {
       data[11] = valueLSB;
     }
 
+    super.send();
+  }
+}
+
+class PitchBendMessage extends MidiMessage {
+  final int channel;
+  final double bend;
+
+  /// Create Pitch Bend Message with a bend value range of -1.0 to 1.0 (default: 0.0).
+  PitchBendMessage({this.channel = 0, this.bend = 0});
+
+  @override
+  void send() {
+    double clampedBend = (bend.clamp(-1, 1) + 1) / 2.0;
+    int targetValue = (clampedBend * 0x3FFF).round();
+
+    int bendMSB = targetValue >> 7;
+    int bendLSB = targetValue & 0x7F;
+
+    data = Uint8List(3);
+    data[0] = 0xE0 + channel;
+    data[1] = bendLSB;
+    data[2] = bendMSB;
+    super.send();
+  }
+}
+
+class PolyATMessage extends MidiMessage {
+  int channel = 0;
+  int note = 0;
+  int pressure = 0;
+
+  /// Create a Polyphonic Aftertouch Message for a single note
+  PolyATMessage({this.channel = 0, this.note = 0, this.pressure = 0});
+
+  @override
+  void send() {
+    data = Uint8List(3);
+    data[0] = 0xA0 + channel;
+    data[1] = note;
+    data[2] = pressure;
+    super.send();
+  }
+}
+
+class ATMessage extends MidiMessage {
+  int channel = 0;
+  int pressure = 0;
+
+  /// Create an Aftertouch Message for a single channel
+  ATMessage({this.channel = 0, this.pressure = 0});
+
+  @override
+  void send() {
+    data = Uint8List(2);
+    data[0] = 0xD0 + channel;
+    data[1] = pressure;
     super.send();
   }
 }